+++ Firestore/core/src/api/firestore.h
@@ -107,7 +107,8 @@ class Firestore : public std::enable_shared_from_this<Firestore> {
   void EnableNetwork(util::StatusCallback callback);
   void DisableNetwork(util::StatusCallback callback);
 
-  void SetIndexConfiguration(std::string config, util::StatusCallback callback);
+  void SetIndexConfiguration(const std::string& config,
+                             const util::StatusCallback& callback);
 
   std::shared_ptr<api::LoadBundleTask> LoadBundle(
       std::unique_ptr<util::ByteStream> bundle_data);
