+++ Firestore/core/src/bundle/bundle_serializer.cc
@@ -45,7 +45,6 @@
 namespace firebase {
 namespace firestore {
 namespace bundle {
-namespace {
 
 using absl::Time;
 using core::Bound;
@@ -73,21 +72,16 @@ using nanopb::Message;
 using nanopb::SetRepeatedField;
 using nanopb::SharedMessage;
 using nlohmann::json;
+using util::JsonReader;
 using util::NoDestructor;
 using util::StatusOr;
 using util::StringFormat;
 using Operator = FieldFilter::Operator;
 
 namespace {
+
 const NoDestructor<Bound> kDefaultBound{Bound::FromValue(
     MakeSharedMessage<google_firestore_v1_ArrayValue>({}), false)};
-}  // namespace
-
-template <typename T>
-const std::vector<T>& EmptyVector() {
-  static NoDestructor<std::vector<T>> empty;
-  return *empty;
-}
 
 Timestamp DecodeTimestamp(JsonReader& reader, const json& version) {
   StatusOr<Timestamp> decoded;
@@ -325,180 +319,6 @@ pb_bytes_array_t* DecodeBytesValue(JsonReader& reader,
 }
 
 }  // namespace
-
-// Mark: JsonReader
-
-const std::string& JsonReader::RequiredString(const char* name,
-                                              const json& json_object) {
-  if (json_object.contains(name)) {
-    const json& child = json_object.at(name);
-    if (child.is_string()) {
-      return child.get_ref<const std::string&>();
-    }
-  }
-
-  Fail("'%s' is missing or is not a string", name);
-  return util::EmptyString();
-}
-
-const std::string& JsonReader::OptionalString(
-    const char* name,
-    const json& json_object,
-    const std::string& default_value) {
-  if (json_object.contains(name)) {
-    const json& child = json_object.at(name);
-    if (child.is_string()) {
-      return child.get_ref<const std::string&>();
-    }
-  }
-
-  return default_value;
-}
-
-const std::vector<json>& JsonReader::RequiredArray(const char* name,
-                                                   const json& json_object) {
-  if (json_object.contains(name)) {
-    const json& child = json_object.at(name);
-    if (child.is_array()) {
-      return child.get_ref<const std::vector<json>&>();
-    }
-  }
-
-  Fail("'%s' is missing or is not an array", name);
-  return EmptyVector<json>();
-}
-
-const std::vector<json>& JsonReader::OptionalArray(
-    const char* name,
-    const json& json_object,
-    const std::vector<json>& default_value) {
-  if (!json_object.contains(name)) {
-    return default_value;
-  }
-
-  const json& child = json_object.at(name);
-  if (child.is_array()) {
-    return child.get_ref<const std::vector<json>&>();
-  } else {
-    Fail("'%s' is not an array", name);
-    return EmptyVector<json>();
-  }
-}
-
-bool JsonReader::OptionalBool(const char* name,
-                              const json& json_object,
-                              bool default_value) {
-  return (json_object.contains(name) && json_object.at(name).is_boolean() &&
-          json_object.at(name).get<bool>()) ||
-         default_value;
-}
-
-const nlohmann::json& JsonReader::RequiredObject(const char* child_name,
-                                                 const json& json_object) {
-  if (!json_object.contains(child_name)) {
-    Fail("Missing child '%s'", child_name);
-    return json_object;
-  }
-  return json_object.at(child_name);
-}
-
-const nlohmann::json& JsonReader::OptionalObject(
-    const char* child_name,
-    const json& json_object,
-    const nlohmann::json& default_value) {
-  if (json_object.contains(child_name)) {
-    return json_object.at(child_name);
-  }
-  return default_value;
-}
-
-double JsonReader::RequiredDouble(const char* name, const json& json_object) {
-  if (json_object.contains(name)) {
-    double result = DecodeDouble(json_object.at(name));
-    if (ok()) {
-      return result;
-    }
-  }
-
-  Fail("'%s' is missing or is not a double", name);
-  return 0.0;
-}
-
-double JsonReader::OptionalDouble(const char* name,
-                                  const json& json_object,
-                                  double default_value) {
-  if (json_object.contains(name)) {
-    double result = DecodeDouble(json_object.at(name));
-    if (ok()) {
-      return result;
-    }
-  }
-
-  return default_value;
-}
-
-double JsonReader::DecodeDouble(const nlohmann::json& value) {
-  if (value.is_number()) {
-    return value.get<double>();
-  }
-
-  double result = 0;
-  if (value.is_string()) {
-    const auto& s = value.get_ref<const std::string&>();
-    auto ok = absl::SimpleAtod(s, &result);
-    if (!ok) {
-      Fail("Failed to parse into double: " + s);
-    }
-  }
-  return result;
-}
-
-template <typename IntType>
-IntType ParseInt(const json& value, JsonReader& reader) {
-  if (value.is_number_integer()) {
-    return value.get<IntType>();
-  }
-
-  IntType result = 0;
-  if (value.is_string()) {
-    const auto& s = value.get_ref<const std::string&>();
-    auto ok = absl::SimpleAtoi<IntType>(s, &result);
-    if (!ok) {
-      reader.Fail("Failed to parse into integer: " + s);
-      return 0;
-    }
-
-    return result;
-  }
-
-  reader.Fail("Only integer and string can be parsed into int type");
-  return 0;
-}
-
-template <typename IntType>
-IntType JsonReader::RequiredInt(const char* name, const json& json_object) {
-  if (!json_object.contains(name)) {
-    Fail("'%s' is missing or is not a double", name);
-    return 0;
-  }
-
-  const json& value = json_object.at(name);
-  return ParseInt<IntType>(value, *this);
-}
-
-template <typename IntType>
-IntType JsonReader::OptionalInt(const char* name,
-                                const json& json_object,
-                                IntType default_value) {
-  if (!json_object.contains(name)) {
-    return default_value;
-  }
-
-  const json& value = json_object.at(name);
-  return ParseInt<IntType>(value, *this);
-}
-
-// Mark: BundleSerializer
 
 BundleMetadata BundleSerializer::DecodeBundleMetadata(
     JsonReader& reader, const json& metadata) const {
